rag_engine/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
fastapi==0.115.0
pydantic==2.10.0
langchain-ollama==0.2.3
google-generativeai==0.6.0
python-dotenv==1.0.1
python-multipart==0.0.6
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssselect2==0.8.0
et_xmlfile==2.0.0
faiss-cpu==1.13.2
filelock==3.20.2
fonttools==4.61.0
fsspec==2025.12.0
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.3
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
pydantic_core==2.27.0
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
reportlab==4.4.6
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
six==1.17.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
weasyprint==67.0
webencodings==0.5.1
xlrd==2.0.2
xxhash==3.6.0
zopfli==0.4.0
zstandard==0.25.0
